+++ UVtools.GUI/Properties/Settings.Designer.cs
@@ -613,7 +613,7 @@ namespace UVtools.GUI.Properties {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("1")]
+        [global::System.Configuration.DefaultSettingValueAttribute("0")]
         public byte LayerRepairDefaultOpeningIterations {
             get {
                 return ((byte)(this["LayerRepairDefaultOpeningIterations"]));